refactor: separando page.tsx em 3 responsabilidades diferentes em hooks
