+++ components/scene/Apartamento0(paraPossivelDebug).jsx
